+++ hammurabi/__init__.py
@@ -71,4 +71,4 @@ try:
 except ImportError:
     pass
 
-__version__ = "0.9.1"
+__version__ = "0.10.0-beta"