+++ CHANGELOG.rst
@@ -53,7 +53,7 @@ Changed
 * Replace click based CLI with a Typer based one
 * Use `__version__` for local documentation generation
 * Update CONTRIBUTING.md regarding documentation config version bump
-* Include `main.py` in test reports
+* Include `main.py` in test reports and add tests
 
 Removed
 ~~~~~~~