+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 
 .. Hyperlinks for releases
 
-.. _Unreleased: https://github.com/gabor-boros/hammurabi/compare/v0.9.1...master
+.. _Unreleased: https://github.com/gabor-boros/hammurabi/compare/v0.10.0-beta...master
 .. _0.1.0: https://github.com/gabor-boros/hammurabi/releases/tag/v0.1.0
 .. _0.1.1: https://github.com/gabor-boros/hammurabi/releases/tag/v0.1.1
 .. _0.1.2: https://github.com/gabor-boros/hammurabi/releases/tag/v0.1.2
@@ -30,9 +30,13 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 .. _0.8.2: https://github.com/gabor-boros/hammurabi/releases/tag/v0.8.2
 .. _0.9.0: https://github.com/gabor-boros/hammurabi/releases/tag/v0.9.0
 .. _0.9.1: https://github.com/gabor-boros/hammurabi/releases/tag/v0.9.1
+.. _0.10.0-beta: https://github.com/gabor-boros/hammurabi/releases/tag/v0.10.0-beta
 
 Unreleased_
 -----------
+
+0.10.0-beta_ - 2020-08-13
+-------------------------
 
 Added
 ~~~~~